refactor(struct): extract EPUB visual fallback tests
